Add PathName type
I noticed several instances of path variables and parameters in the repository that are declared as weak types. Due to the special properties of paths, I opted to implement a stronger type that throws whenever an illegal path is passed to its constructor. Checking legality is made simple by creating a dummy ifstream object and testing whether the stream is valid.