Find and update weak path types to PathName
Using CodeCompass and Visual Studio, I searched for path name parameters and variables that could be prime candidates for the stronger type I implemented in the previous commit. In this commit, I changed char pointer types to PathName, updating the corresponding pieces of code, paying attention to virtual functions and circular dependencies. It is important to note that althogh the current state of the repository compiles without error, there are still weak path types remaining in the code.